geo_app/requirements.txt
streamlit>=1.32.0
openai>=1.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
nltk>=3.8.0
textstat>=0.7.0
